requirements.txt
Flask
SpeechRecognition
deep-translator
pydub
PyPDF2
fpdf2
python-docx
